refactor(LocalUserAccountProvider): Separate LocalUserAccountProvider into abstract base class and h
