+++ packages/neo-one-client-core/src/errors.ts
@@ -154,3 +154,8 @@ export const InvalidMasterPathError = makeErrorWithCode(
   'INVALID_HD_MASTER_PATH',
   (path: readonly number[]) => `Invalid masterPath returned by HDStore: ${path}`,
 );
+
+export const NotImplementedError = makeErrorWithCode(
+  'NOT_IMPLEMENTED',
+  (name: string) => `${name} not implemented for this provider.`,
+);
